+++ src/main/java/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/type/index/CompositeIndex.java
@@ -14,7 +14,8 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
  * If index with provided name exist, but with different type, index will be re-created.
  * If existing index build on different properties, error will be thrown.
  * <p>Use {@link CompositeIndex.List} to define more than one index.</p>
- * <p>Annotation doesn't cover all possible index options.</p>
+ * <p>Annotation doesn't cover all possible index options. Use specific annotation for fulltext lucene index:
+ * {@link ru.vyarus.guice.persist.orient.db.scheme.initializer.ext.type.index.lucene.CompositeLuceneIndex}.</p>
  *
  * @author Vyacheslav Rusakov
  * @see <a href="http://www.orientechnologies.com/docs/last/orientdb.wiki/Indexes.html">docs</a>