+++ src/main/java/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/type/index/IndexTypeExtension.java
@@ -15,6 +15,8 @@ import ru.vyarus.guice.persist.orient.db.scheme.initializer.ext.field.index.Inde
 import javax.inject.Singleton;
 
 /**
+ * {@link CompositeIndex} scheme model type extension.
+ *
  * @author Vyacheslav Rusakov
  * @since 09.03.2015
  */