+++ src/main/java/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/type/index/lucene/MultipleLuceneIndexesTypeExtension.java
@@ -1,0 +1,41 @@
+package ru.vyarus.guice.persist.orient.db.scheme.initializer.ext.type.index.lucene;
+
+import com.orientechnologies.orient.object.db.OObjectDatabaseTx;
+import ru.vyarus.guice.persist.orient.db.scheme.initializer.core.spi.SchemeDescriptor;
+import ru.vyarus.guice.persist.orient.db.scheme.initializer.core.spi.type.TypeExtension;
+
+import javax.inject.Inject;
+import javax.inject.Singleton;
+
+/**
+ * {@link CompositeLuceneIndex.List} scheme model type extension.
+ *
+ * @author Vyacheslav Rusakov
+ * @since 20.06.2015
+ */
+@Singleton
+public class MultipleLuceneIndexesTypeExtension implements TypeExtension<CompositeLuceneIndex.List> {
+
+    private final LuceneIndexTypeExtension extension;
+
+    @Inject
+    public MultipleLuceneIndexesTypeExtension(final LuceneIndexTypeExtension extension) {
+        this.extension = extension;
+    }
+
+    @Override
+    public void beforeRegistration(final OObjectDatabaseTx db, final SchemeDescriptor descriptor,
+                                   final CompositeLuceneIndex.List annotation) {
+        for (CompositeLuceneIndex index : annotation.value()) {
+            extension.beforeRegistration(db, descriptor, index);
+        }
+    }
+
+    @Override
+    public void afterRegistration(final OObjectDatabaseTx db, final SchemeDescriptor descriptor,
+                                  final CompositeLuceneIndex.List annotation) {
+        for (CompositeLuceneIndex index : annotation.value()) {
+            extension.afterRegistration(db, descriptor, index);
+        }
+    }
+}