+++ src/test/groovy/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/type/index/lucene/MultipleLuceneIndexesRemoteTest.groovy
@@ -1,0 +1,11 @@
+package ru.vyarus.guice.persist.orient.db.scheme.initializer.ext.type.index.lucene
+
+import ru.vyarus.guice.persist.orient.util.remoteext.UseRemote
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 20.06.2015
+ */
+@UseRemote
+class MultipleLuceneIndexesRemoteTest extends CompositeLuceneIndexTest {
+}